refactor input utility functions into util module
